Refactor: move cached session to repondeur
